requirements.txt
speedtest-cli
requests
matplotlib
numpy
